move Digitalbonus from /e-rechnung to /digitalbonus, cover both e-invoice and AI consulting
E-invoice is technical docs only. Digitalbonus is a cross-cutting funding page
for both business areas. Added Digitalbonus teaser to AI consulting page.
